requirements.txt
streamlit
pandas
plotly
mysql-connector-python
python-dotenv

# BigQuery e autenticação
google-cloud-bigquery
google-auth
google-auth-oauthlib
pyarrow
